pyproject.toml
[tool.poetry]
name = "blum-tap-clicker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.0.6"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
mss = "^9.0.1"
pyautogui = "^0.9.54"
opencv-python = "^4.10.0.82"
keyboard = "^0.13.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
